refactor: move error classes to where they're raised
I think "by domain" is a better way of grouping them than stuffing
unrelated error classes in a file together. This is an implementation
detail anyway, given that they're imported to the top level in
__init__.py.

Also add "make fix" to run "ruff check --fix".
